move router to features
