+++ src/components/table/DefaultFileCard.tsx
@@ -13,7 +13,7 @@ import { Dialog, DialogContent } from "../ui/dialog"
 import UpdatesDialog from "./UpdatesDialog"
 import { isFileLinkTracked, isUrlTracked } from "@src/content/versionControl"
 import { Tooltip, TooltipContent, TooltipTrigger } from "../ui/tooltip"
-import Link from "../router/Link"
+import Link from "../../features/router/Link"
 
 export interface DefaultFileCardProps {
     fileLink: FileLink
