+++ src/pages/content/Providers.tsx
@@ -5,7 +5,7 @@ import { CommandProvider } from "@src/components/command/CommandContext"
 import ConfigContextProvider from "@src/contexts/ConfigContext"
 import { TooltipProvider } from "@src/components/ui/tooltip"
 import { PageStateContextProvider } from "@src/contexts/PageStateContext"
-import { SmoothRouterListener } from "@src/components/router/SmoothRouterListener"
+import { SmoothRouterListener } from "@src/features/router/SmoothRouterListener"
 import { PreferredTheme, ThemeProvider } from "@src/contexts/ThemeContext"
 import { Toaster } from "@src/components/ui/toaster"
